requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
cvxpy>=1.2.0
scikit-learn>=1.0.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0
scipy>=1.7.0
networkx>=2.6.0
shapely>=1.8.0
plotly>=5.0.0
kaleido>=0.2.1




